+++ pkg/config/config.go
@@ -65,6 +65,8 @@ type Ingest struct {
 
 type File struct {
 	Filename string
+	Loop     bool
+	Chunks   int
 }
 
 type Aws struct {
